requirements.txt
flask
flask_sqlalchemy
flask-migrate
mysql-connector-python
python-dotenv
gunicorn
Werkzeug
SQLAlchemy
alembic